Update res currency update services to new api.
